* fixed ufo icons for mingw build * fixed compile error in irc.c for vc * redruced some warnings (int => size_t) * some fixme comments for server.mk

+++ src/qcommon/files.c
@@ -1307,7 +1307,7 @@ int FS_Write(const void *buffer, int len, qFILE * f)
 /**
  * @brief
  */
-int FS_WriteFile(const void *buffer, int len, const char *filename)
+int FS_WriteFile(const void *buffer, size_t len, const char *filename)
 {
 	qFILE f;
 	int c, lencheck;
